refactor: leverage helpers
