+++ packages/generators/src/init-generator.ts
@@ -1,12 +1,11 @@
 import { blue, yellow } from "colorette";
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from "fs";
 import path from "path";
-import * as Question from "./utils/scaffold-utils";
 
 import { CustomGenerator } from "./types";
-import { existsSync, mkdirSync } from "fs";
+import { getInstaller, getTemplate } from "./utils/helpers";
+import * as Question from "./utils/scaffold-utils";
 import handlers from "./handlers";
-
-import { readFileSync, writeFileSync } from "fs";
 
 /**
  *
@@ -62,22 +61,7 @@ export default class InitGenerator extends CustomGenerator {
             }
         }
 
-        if (!this.supportedTemplates.includes(this.template)) {
-            this.utils.logger.warn(
-                `⚠ ${this.template} is not a valid template, please select one from below`,
-            );
-
-            const { selectedTemplate } = await Question.List(
-                this,
-                "selectedTemplate",
-                "Select a valid template from below:",
-                this.supportedTemplates,
-                "default",
-                false,
-            );
-
-            this.template = selectedTemplate;
-        }
+        this.template = await getTemplate.call(this);
 
         await handlers[this.template].questions(this, Question);
 
@@ -101,23 +85,7 @@ export default class InitGenerator extends CustomGenerator {
     }
 
     public async installPlugins(): Promise<void> {
-        const installers = this.utils.getAvailableInstallers();
-
-        if (installers.length === 1) {
-            [this.packageManager] = installers;
-        } else {
-            // Prompt for the package manager of choice
-            const defaultPackager = this.utils.getPackageManager();
-            const { packager } = await Question.List(
-                this,
-                "packager",
-                "Pick a package manager:",
-                installers,
-                defaultPackager,
-                this.force,
-            );
-            this.packageManager = packager;
-        }
+        this.packageManager = await getInstaller.call(this);
 
         const opts: {
             dev?: boolean;
